backend/requirements.txt
blinker==1.8.2
click==8.1.7
exceptiongroup==1.2.1
flask==3.0.3
Flask-Cors==4.0.1
importlib-metadata==7.2.0
iniconfig==2.0.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pluggy==1.5.0
prometheus-client==0.11.0
pytest==8.2.2
tomli==2.0.1
werkzeug==3.0.3
zipp==3.19.2
psycopg2-binary==2.9.3
